refactor: move and test the ProductCount
